+++ CHANGELOG.md
@@ -10,6 +10,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) 
   `anyCharacterOf(String characters)` and `anyCharacterExcept(String characters)`.
 - Added methods for word characters and boundaries: `wordCharacter()`, `nonWordCharacter()`, `wordBoundary()` and `nonWordBoundary()`.
   Also, added the method `word()` which searches for words.
+- All greedy quantifiers are now available.
+- Added methods using dot (`.`): `anyCharacter()` and `anything()`.
+- It is now possible to make existing quantifiers reluctant or possessive by appending the method `reluctant()` or `possessive()`
+  after the quantifier.
 
 ### Changed
 - Spotbugs annotations is removed as dependency.